Initial commit with restaurant_basic
This initial commit contains the web server restaurant_basic.
restaurant_basic show cases python3, CRUD, SQLAlchemy.

restaurant_basic allows a user to create, edit, delete and list
restaurants in a database.  There are security holes and poor
practices in this commit that will be fixed in later commits.
Please do not use this commit as an example for production.