Udacity's flask_basic folder initial commit
This example Restaurant web server is from Udacity's
github project.  The URL is at the bottom of this commit message.
A database is included in the commit.

Technologies used:
Python, Flask, SQLAlchemy, HTML templates, routing.

Over time I plan to update, evolve and fix the bugs in this
initial commit.

https://github.com/udacity/Full-Stack-Foundations/tree/master/Lesson-3